Extract ClickableTitleBar into separate files
